Nueva actualizacion de eliminacion de cuentas
Nueva actualizacion de eliminacion de cuentas

+++ Economia/modelo/eliminarCuenta.php
@@ -5,20 +5,33 @@ $cod_grupo = $_POST['cod_grupo'];
 $cod_bloque = $_POST['cod_bloque'];
 $cod_rubro = $_POST['cod_rubro'];
 $cod_cuenta = $_POST['cod_cuenta'];
-$cod_rubro = '0'.$cod_rubro;
-echo $cod_grupo.".". $cod_bloque.".". $cod_rubro.".". $cod_cuenta;
 
-//validar saldo 0
+//echo $cod_grupo . "." . $cod_bloque . "." . $cod_rubro . "." . $cod_cuenta;
 
-$queryD = "DELETE FROM cuenta WHERE cod_cuenta='$cod_cuenta' and cod_rubro='$cod_rubro' and cod_bloque='$cod_bloque' and cod_grupo='$cod_grupo'";
-echo $queryD;
-$rsD = mysqli_query($conex, $queryD);
+$queryC = "SELECT saldo FROM cuenta WHERE cod_rubro = '$cod_rubro' and cod_bloque= '$cod_bloque' and cod_grupo= '$cod_grupo' and cod_cuenta='$cod_cuenta'";
+$rsC = mysqli_query($conex, $queryC);
+$rowC = mysqli_fetch_assoc($rsC);
 
-if ($rsD) {
-    echo "<script>alert('Cuenta Eliminada')</script>";
-}else{
-    echo "<script>alert('error')</script>";
-}
+$saldo = $rowC['saldo'];
 
+if ($saldo == 0) {
+
+    //----------------------------------------------------------
+    $queryD = "DELETE FROM cuenta WHERE cod_cuenta='$cod_cuenta' and cod_rubro='$cod_rubro' and cod_bloque='$cod_bloque' and cod_grupo='$cod_grupo'";
+    //echo $queryD;
+    $rsD = mysqli_query($conex, $queryD);
 ?>
-<script src="https://code.jquery.com/jquery-latest.js"></script>
+<script>Swal.fire('Cuenta eliminada con éxito', '', 'success')</script>
+<?php
+    /* if ($rsD) {
+        echo "";
+    } else {
+        echo "<script>Swal.fire('Error al eliminar', '', 'error')</script>";
+    } */
+    //----------------------------------------------------------
+
+}else {
+    echo "<script>Swal.fire('Esta cuenta no puede ser eliminada. Verificar saldo', '', 'error')</script>";
+}
+echo '<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>';
+?>
